requirements.txt

# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.10.0
olefile>=0.47

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# NLP
kss>=6.0.0

# Visualization & Dashboard
streamlit>=1.30.0
plotly>=5.18.0
matplotlib>=3.8.0

# Statistical Modeling
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Phase 1 New Additions (2026-02-09)
selenium>=4.10.0              # Web Scraping
# ekonlpy>=2.0.0              # Korean NLP (Optional - Build issues on Windows)
wordcloud>=1.9.2              # Word Cloud
Pillow>=10.0.0                # Image Processing
openpyxl>=3.1.0               # Excel Handling
PyYAML>=6.0
